Got rid of console.log in index.html Moved InstallButton function to separate component Moved Pwa component to hooks  Moved BeforeInstall Interface to models Fixed styling of Button for PWA
