+++ public/index.html
@@ -30,16 +30,6 @@
 
             gtag('config', 'UA-155742060-1');
         </script>
-
-        <script>
-            window.addEventListener('beforeinstallprompt', function(e) {
-                console.log(e);
-                console.log(e.platforms);
-                e.userChoice.then(function(choiceResult) {
-                    console.log(choiceResult.outcome);
-                }, handleError);
-            });
-        </script>
     </head>
     <body>
         <noscript>You need to enable JavaScript to run this app.</noscript>
