+++ src/app/models/BeforeInstall.ts
@@ -1,0 +1,8 @@
+export interface BeforeInstallPromptEvent extends Event {
+    readonly platforms?: string[];
+    readonly userChoice?: Promise<{
+        outcome: 'accepted' | 'dismissed';
+        platform: string;
+    }>;
+    prompt?: () => Promise<void>;
+}
